Change AfkRewardRate rate to float instead of integer. Also use that field as hours, instead of minutes. Change the afk rewards calculation accordingly

+++ apps/game_backend/lib/game_backend/campaigns/rewards/afk_reward_rate.ex
@@ -12,8 +12,8 @@ defmodule GameBackend.Campaigns.Rewards.AfkRewardRate do
   schema "afk_reward_rates" do
     belongs_to(:user, User)
     belongs_to(:currency, Currency)
-    # Per minute
-    field(:rate, :integer)
+    # Per hour
+    field(:rate, :float)
 
     timestamps()
   end
